Improve M
The capital M is once again missing its top-left corner pixel. The
missing pixel isn't purely for softness; it's supposed to create jaggies
that convey the "shoulders" of the letter.

Whether that works is somewhat debatable. What isn't debatable is that
the blocky capital M looked stiff and out of place, particularly after
so many other letters were softened.